* Fix some gcc warnings in DxSound.cpp (not that it's compileable in mingw yet...)
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2643 37977431-3df6-0310-b722-df95706aa16b

+++ rts/System/Platform/Win/DxSound.cpp
@@ -210,7 +210,7 @@ void CDxSound::PlaySample(int id,float volume)
 		return;
 	}
 	
-	buffers[num]->SetVolume(DSBVOLUME_MIN*v);
+	buffers[num]->SetVolume(int(DSBVOLUME_MIN*v));
 	buffers[num]->SetPan(0);
 //	if(reset)
 //		buffers[num]->SetCurrentPosition( 0L );    
@@ -260,8 +260,8 @@ void CDxSound::PlaySample(int id,const float3& p,float volume)
 	if( FAILED( hr = RestoreBuffers(num) ) )
 		return;
 	
-	buffers[num]->SetVolume(DSBVOLUME_MIN*v-100);
-	buffers[num]->SetPan(pan);
+	buffers[num]->SetVolume(int(DSBVOLUME_MIN*v-100));
+	buffers[num]->SetPan(int(pan));
 
 	if( FAILED( hr = buffers[num]->Play( 0, 0, 0/*dwLooped*/ ) ) ){
